Cardano.Kuber.Api

TxBuilder and companion functions

Transaction Builder Object

data TxBuilder

TxBuilder is not to be directly used but, parts of it are constructed using helper functions. Multiple builder parts can be combined to construct full transaction specification

Transaction Inputs

Add the txin to transaction input

txConsumeTxIn :: TxIn -> TxBuilder

txConsumeUtxo :: TxIn -> TxOut CtxUTxO BabbageEra -> TxBuilder

txConsumeUtxos :: UTxO BabbageEra -> TxBuilder

txRedeemTxin :: IsPlutusScript sc => TxIn -> sc -> ScriptData -> Maybe ExecutionUnits -> TxBuilder

Add a script utxo-reference containing inline-datum to transaction input. Script code and Reedemer should be passed.

txRedeemUtxo :: IsPlutusScript sc => TxIn -> TxOut CtxUTxO BabbageEra -> sc -> ScriptData -> Maybe ExecutionUnits -> TxBuilder

Add a script utxo containing inline-datum to transaction input. Script code and redeemer should be passed for building transaction.

txRedeemUtxoWithDatum :: IsPlutusScript sc => TxIn -> TxOut CtxUTxO BabbageEra -> sc -> ScriptData -> ScriptData -> Maybe ExecutionUnits -> TxBuilder

Add a script utxo containing datum-hash to transaction input . Script code, datum matching datumHash and redeemer should be passed for building transaction.

txRedeemUtxoWithReferenceScript :: ScriptReferenceTxIn -> TxIn -> TxOut CtxUTxO BabbageEra -> ScriptData -> Maybe ExecutionUnits -> TxBuilder

Add a script utxo containing inline-datum to transaction input. Script code is inlined in provided TransactionInput. The script reference input will be automatically added to transaction reference inputs.

txRedeemTxinWithReferenceScript :: ScriptReferenceTxIn -> TxIn -> TxOut CtxUTxO BabbageEra -> ScriptData -> Maybe ExecutionUnits -> TxBuilder

Add a script txIn containing inline-datum to transaction input. Script code is inlined in provided TransactionInput. The script reference input will be automatically added to transaction reference inputs.

txRedeemUtxoWithDatumAndReferenceScript :: ScriptReferenceTxIn -> TxIn -> TxOut CtxUTxO BabbageEra -> ScriptData -> ScriptData -> Maybe ExecutionUnits -> TxBuilder

Add a script utxo txin containing datum-hash to transaction input. Script code is inlined in provided TransactionInput. The script reference input will be automatically added to transaction reference inputs.

txRedeemTxinWithDatumAndReferenceScript :: ScriptReferenceTxIn -> TxIn -> TxOut CtxUTxO BabbageEra -> ScriptData -> ScriptData -> Maybe ExecutionUnits -> TxBuilder

Add a script utxo containing datum-hash to transaction input. Script code is inlined in provided TransactionInput. The script reference input will be automatically added to transaction reference inputs.

TxSelections

TxSelections represent Utxos that can be used for balancing a transaction. They will be added to input if required for balancing the transaciton. Such utxos should be related to user's wallet and

txWalletAddress :: AddressInEra BabbageEra -> TxBuilder

txWalletAddresses :: [AddressInEra BabbageEra] -> TxBuilder

txWalletUtxos :: UTxO BabbageEra -> TxBuilder

txWalletUtxo :: TxIn -> TxOut CtxUTxO BabbageEra -> TxBuilder

txWalletSignKey :: SigningKey PaymentKey -> TxBuilder

add signkey to the selection. All the utxos from the enterprise address of the signkey will be automatically added to selection too. The TxBuilder containing signkeys when built to transaction, transaction will contain signatures if their utxo is used.

In order to automatically sign tx for address other than enterprise address, both address and signkey must be present in the builder.

txWalletSignKeys :: [SigningKey PaymentKey] -> TxBuilder

Payment functions

txPayTo :: AddressInEra BabbageEra -> Value -> TxBuilder

Pay to this address in transaction

txPayToPkh :: PubKeyHash -> Value -> TxBuilder

Pay to the enterprise address of this PublicKeyHash

txPayToScript :: AddressInEra BabbageEra -> Value -> Hash ScriptData -> TxBuilder

Pay to script address with datumHash

txPayToScriptWithData :: AddressInEra BabbageEra -> Value -> ScriptData -> TxBuilder

Pay to script address and inline the datum in utxo

txPayToScriptWithReference :: Script -> Value -> Hash ScriptData -> TxBuilder

Pay to the script and inline it in the utxo. Script enterprise address is derrived from script hash

txPayToScriptWithDataAndReference :: Script -> Value -> ScriptData -> TxBuilder

Pay to script with inline both datum and inline it in datum. Script enterprise address is derrived from script hash

txPayToWithReferenceScript :: AddressInEra BabbageEra -> Value -> Script -> TxBuilder

Pay to address and inline the script in resulting utxo.

Minting functions

txMintSimpleScript :: IsSimpleScript script => script -> [(AssetName, Quantity)] -> TxBuilder

Mint token with simple script

txMintPlutusScript :: IsPlutusScript script => script -> ScriptData -> [(AssetName, Quantity)] -> TxBuilder

Mint token with plutus v1 or v2 script

Reference Inputs

txReferenceTxIn :: TxIn -> TxBuilder

txReferenctUtxo :: TxIn -> TxOut CtxUTxO BabbageEra -> TxBuilder

Extra Vkey Witnesses in Transaction

txSignBy :: AddressInEra BabbageEra -> TxBuilder

Mark this address as txExtraKeyWitness in the transaction object.

txSignByPkh :: PubKeyHash -> TxBuilder

Mark this PublicKeyhash as txExtraKeyWitness in the transaction object.

txSign :: SigningKey PaymentKey -> TxBuilder

Explicitly set transaction fee

txSetFee :: Integer -> TxBuilder

explicitly set Fee for the transaction

Change address to send unbalanced value to. If change address is missing, it is automatically selected from one of the wallet addresses.

txChangeAddress :: AddressInEra BabbageEra -> TxBuilder

Transactoin validity functions

txValidFromPosixTime :: POSIXTime -> TxBuilder

txValidUntilPosixTime :: POSIXTime -> TxBuilder

txValidPosixTimeRange :: POSIXTime -> POSIXTime -> TxBuilder

txValidFromSlot :: SlotNo -> TxBuilder

txValidUntilSlot :: SlotNo -> TxBuilder

txValidSlotRange :: SlotNo -> SlotNo -> TxBuilder

Transactoin validity functions

these functions help dealing with plutus validator and cardano-api script types.

class IsPlutusScript sc

Minimal complete definition

toTxPlutusScript

class IsSimpleScript sc

Minimal complete definition

toTxSimpleScript

data TxScript

Constructors

TxScriptSimple TxSimpleScript

TxScriptPlutus TxPlutusScript

data TxPlutusScript

Constructors

TxPlutusScriptV1 (PlutusScript PlutusScriptV1)

TxPlutusScriptV2 (PlutusScript PlutusScriptV2)

data TxSimpleScript

Constructors

TxSimpleScriptV1 (SimpleScript SimpleScriptV1)

TxSimpleScriptV2 (SimpleScript SimpleScriptV2)

toTxSimpleScript :: IsSimpleScript sc => sc -> TxSimpleScript

toTxPlutusScript :: IsPlutusScript sc => sc -> TxPlutusScript

hashPlutusScript :: TxPlutusScript -> ScriptHash

plutusScriptAddr :: TxPlutusScript -> NetworkId -> AddressInEra BabbageEra

plutusScriptToScriptAny :: TxPlutusScript -> ScriptInAnyLang

txScriptPolicyId :: TxScript -> PolicyId

txScriptAddress :: TxScript -> NetworkId -> StakeAddressReference -> AddressInEra BabbageEra

txScriptHash :: TxScript -> ScriptHash

hashTxScript :: TxScript -> ScriptHash

txScriptToScriptAny :: TxScript -> ScriptInAnyLang

Error Class

data ErrorType

Constructors

ConnectionError

BalancingError

InsufficientInput

EraMisMatch

NodeQueryError

LibraryError

ParserError

PlutusScriptError

PlutusExecutionError

ExUnitCalculationError

FeatureNotSupported

TxValidationError

BadMetadata

TxSubmissionError

WrongScriptType

data FrameworkError

Constructors

FrameworkError

Fields

feType :: ErrorType

feMessage :: String

FrameworkErrors [FrameworkError]

throwFrameworkError :: Applicative m => Either FrameworkError a -> m a

Kontract and ChainApi related API

data Exception e => Kontract api w e r

Constructors

KResult !r

KError !e

KLift (api -> IO (Either e r))

kError :: ErrorType -> String -> Kontract api w FrameworkError r

kWrapParser :: Either String r -> Kontract api w FrameworkError r

kGetBackend :: Exception e => Kontract a w e a

evaluateKontract :: Exception e => a -> Kontract a w e r -> IO (Either e r)

eitherToKontract :: Exception e => Either e r -> Kontract api w e r

class HasChainQueryAPI a where

Methods

kGetNetworkId :: Kontract a w FrameworkError NetworkId

kQueryProtocolParams :: Kontract a w FrameworkError ProtocolParameters

kQuerySystemStart :: Kontract a w FrameworkError SystemStart

kQueryGenesisParams :: Kontract a w FrameworkError GenesisParameters

kQueryUtxoByAddress :: Set AddressAny -> Kontract a w FrameworkError (UTxO BabbageEra)

kQueryUtxoByTxin :: Set TxIn -> Kontract a w FrameworkError (UTxO BabbageEra)

kQueryChainPoint :: Kontract a w FrameworkError ChainPoint

class HasSubmitApi a where

Methods

kSubmitTx :: Tx BabbageEra -> Kontract a w FrameworkError ()

class HasKuberAPI a where

Methods

kTxBuildTxBody :: TxBuilder -> Kontract a w FrameworkError (TxBody BabbageEra)

kBuildTx :: TxBuilder -> Kontract a w FrameworkError (Tx BabbageEra)

kTimeToSlot :: POSIXTime -> Kontract a w FrameworkError SlotNo

kSlotToTime :: SlotNo -> Kontract a w FrameworkError POSIXTime

kEvaluateExUnits :: Tx BabbageEra -> Kontract a w FrameworkError (Map ScriptWitnessIndex (Either FrameworkError ExecutionUnits))

kCalculateMinFee :: Tx BabbageEra -> Kontract a w FrameworkError Lovelace

kBuildAndSubmit :: TxBuilder -> Kontract a w FrameworkError (Tx BabbageEra)

class HasLocalNodeAPI a where

Methods

kQueryEraHistory :: Kontract a w FrameworkError (EraHistory CardanoMode)

Chain info Structures

newtype ChainConnectInfo

Constructors

ChainConnectInfo (LocalNodeConnectInfo CardanoMode)

data RemoteKuberConnection

createRemoteKuberConnection :: NetworkId -> UrlString -> Maybe ApiKeyString -> IO RemoteKuberConnection

createRemoteKuberConnection' :: NetworkId -> BaseUrl -> Maybe ApiKeyString -> IO RemoteKuberConnection

chainInfoFromEnv :: IO ChainConnectInfo

Using Environment variables determine the NETWORK. NETWORK can be "mainnet", "testnet" or "networkMagic number". If CARDANO_NODE_SOCKET_PATH environment variable is set, the socket path is set to it's value Otherwise CARDANO_HOME or "$HOME.cardano" is used and the socket path becomes "$CARDANO_HOMEnode.socket"

chainInfoFromEnv' :: EnvVariable -> IO ChainConnectInfo

Read Network value from the environment variable and then determine connection info If CARDANO_NODE_SOCKET_PATH environment variable is set, the socket path is set to it's value Otherwise CARDANO_HOME or "$HOME.cardano" is used and the socket path becomes "$CARDANO_HOMEnode.socket"

chainInfoMainnet :: IO ChainConnectInfo

chainInfoTestnet :: IO ChainConnectInfo

localNodeConnInfo :: NetworkId -> FilePath -> LocalNodeConnectInfo CardanoMode

getWorkPath :: [FilePath] -> IO FilePath

get absolute path given a directory or file path relative to work directory. the absolute path is "CARANO_HOME/...paths" value to the path

getWorkPathFunc :: IO ([FilePath] -> FilePath)

Get WrokPath calculatin Function getWorkPath function can throw errors. The error is only during initialization So if the function succeeds, it returns pure function to calculate filepath which can be reused.

getNetworkFromEnv :: EnvVariable -> IO (String, NetworkId)

Given environment variable key, read the environmet variable and return network Id. The value maybe network name or network magic.